fix: add variant size for card profile
